Move request.rs into websocket module
